check user to be connected before disconnecting, a little cleaning

+++ src/main/java/com/rahnemacollege/configuration/WebSocketConfig.java
@@ -2,7 +2,6 @@ package com.rahnemacollege.configuration;
 
 import com.rahnemacollege.model.User;
 import com.rahnemacollege.service.BidService;
-import com.rahnemacollege.service.UserDetailsServiceImpl;
 import com.rahnemacollege.service.UserService;
 import com.rahnemacollege.util.JwtTokenUtil;
 import com.rahnemacollege.util.exceptions.MessageException;
@@ -18,7 +17,6 @@ import org.springframework.messaging.support.ChannelInterceptor;
 import org.springframework.messaging.support.MessageHeaderAccessor;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBroker;
 import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
 import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerConfigurer;
@@ -34,13 +32,11 @@ public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
     private JwtTokenUtil tokenUtil;
     private UserService userService;
     private BidService bidService;
-    private UserDetailsService userDetailsService;
 
-    public WebSocketConfig(JwtTokenUtil tokenUtil, UserService userService, BidService bidService, UserDetailsServiceImpl userDetailsService) {
+    public WebSocketConfig(JwtTokenUtil tokenUtil, UserService userService, BidService bidService) {
         this.tokenUtil = tokenUtil;
         this.userService = userService;
         this.bidService = bidService;
-        this.userDetailsService = userDetailsService;
     }
 
     private static Logger logger = org.slf4j.LoggerFactory.getLogger(WebSocketConfig.class);
@@ -65,7 +61,7 @@ public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
             @Override
             public Message<?> preSend(Message<?> message, MessageChannel channel) {
                 StompHeaderAccessor headerAccessor = MessageHeaderAccessor.getAccessor(message, StompHeaderAccessor.class);
-                if (StompCommand.CONNECT.equals(headerAccessor.getCommand())) {
+                if (headerAccessor != null && StompCommand.CONNECT.equals(headerAccessor.getCommand())) {
                     logger.info("try to connect");
                     String jwtToken = Objects.requireNonNull(headerAccessor.getFirstNativeHeader("auth")).substring(7);
                     String id = tokenUtil.getIdFromToken(jwtToken);
@@ -76,12 +72,14 @@ public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
                     headerAccessor.setUser(u);
                     logger.info("the user with session id " + headerAccessor.getSessionId() + " connected");
                 }
-                if (StompCommand.SUBSCRIBE.equals(headerAccessor.getCommand())) {
-                    System.err.println(headerAccessor.getDestination().toString());
+                if (headerAccessor != null && StompCommand.SUBSCRIBE.equals(headerAccessor.getCommand())) {
+                    System.err.println(headerAccessor.getDestination());
                 }
-                if (StompCommand.DISCONNECT.equals(headerAccessor.getCommand())) {
+                if (headerAccessor != null && StompCommand.DISCONNECT.equals(headerAccessor.getCommand()) && headerAccessor.getUser() != null) {
+
                     User user = userService.findUserId(Integer.valueOf(headerAccessor.getUser().getName()));
                     bidService.removeFromAllAuction(user);
+
                 }
                 return message;
             }
